// frontend/src/types.ts
export interface Service {
  id: number;
  name: string;
  type: 'NGO' | 'Police' | 'Helpline' | 'SafeZone';
  description: string;
  address: string;
  lat: number;
  lng: number;
  phone: string;
  email: string;
  hours: string;
  languages: string;
  service_tags?: string | null;
  verified: number;
}